+++ src/@next/components/atoms/PlaceholderImage/stories.tsx
@@ -1,0 +1,8 @@
+import { storiesOf } from "@storybook/react";
+import React from "react";
+
+import { PlaceholderImage } from ".";
+
+storiesOf("@components/atoms/PlaceholderImage", module).add("default", () => (
+  <PlaceholderImage />
+));