+++ src/@next/components/atoms/index.ts
@@ -22,3 +22,4 @@ export * from "./Attribute";
 export * from "./Label";
 export * from "./InputLabel";
 export * from "./DropdownSelect";
+export * from "./PlaceholderImage";