+++ src/@next/components/containers/Money/Money.tsx
@@ -3,10 +3,7 @@ import React from "react";
 import { I18nContext } from "..";
 import { IProps } from "./types";
 
-export const Money: React.StatelessComponent<IProps> = ({
-  money,
-  defaultValue,
-}) => (
+export const Money: React.FC<IProps> = ({ money, defaultValue }: IProps) => (
   <I18nContext.Consumer>
     {({ language }) => {
       if (!money) {