+++ src/checkout/components/CartSummary/Subtotal.tsx
@@ -4,7 +4,7 @@ import { Money } from "@components/containers";
 
 import { Checkout } from "../../../checkout/types/Checkout";
 import { CartLineInterface } from "../../../components/CartProvider/context";
-import { getTotal } from "../../../components/CartProvider/uitls";
+import { getTotal } from "../../../components/CartProvider/utils";
 import { ShopContext } from "../../../components/ShopProvider/context";
 import { maybe } from "../../../core/utils";
 import { VariantList } from "../../../views/Product/types/VariantList";