+++ src/views/Product/View.tsx
@@ -38,7 +38,7 @@ const extractMeta = (product: ProductDetails_product) => ({
     },
   ],
   description: product.seoDescription || product.descriptionJson,
-  image: product.thumbnail.url,
+  image: maybe(() => product.thumbnail.url, null),
   title: product.seoTitle || product.name,
   type: "product.item",
   url: window.location.href,